requirements.txt
Flask>=2.3.0
requests>=2.30.0
nltk>=3.8.0
# newsapi-python>=0.2.7 # Keep if using as fallback
numpy>=1.23.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

news-please==1.6.13 # THIS IS IMPORTANT - it brings newspaper4k
selenium==4.26.1
Unidecode==1.3.8
beautifulsoup4==4.12.3
lxml>=4.9.0
# Other dependencies of news-please and newspaper4k will be pulled automatically

SQLAlchemy==1.4.52
yfinance>=0.2.0
pandas>=1.5.0